sql-generator/requirements.txt
pdfplumber>=0.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=1.5.0
python-dateutil>=2.8.0
extract-msg>=0.41.0
